breakwall_env: fix for collision.
This commit fixes collision with top wall and the bottom wall.
For the top wall it actually needed to making a vertical speed an
absolute value, as ball need to have a vertical positive speed and not
oscilate in the upper corner.
For the bottom wall ball shouldn't bounce at all.